pyproject.toml
[tool.poetry]
name = "hh-vacancy-etl-airflow"
version = "0.1.0"
description = ""
authors = ["Nadezhda_Popova <nadezhdapopova13@yandex.ru>"]

[tool.poetry.dependencies]
python = ">=3.12"
requests = "^2.33.1"
python-dotenv = "^1.2.2"
logging = "^0.4.9.6"
psycopg2 = "^2.9.11"
curl-cffi = "^0.15.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "black (>=26.3.1,<27.0.0)"
]

[tool.black]
line-length = 119
exclude = """
/(
    .git
    .venv
    venv
)/
"""

[tool.isort]
line_length = 119
multi_line_output = 3
